Renamed models folder in classification to avoid namespace collision.
