+++ mycode/classification/classification_models/basic_conv1d.py
